+++ packages/build-app-helpers/src/tests/fixtures/analyzeRuntime/page/store.js
@@ -1,0 +1,2 @@
+import { createStore } from 'ice';
+export default createStore;