+++ packages/build-app-templates/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.1
+
+- [fix] `appConfig` need a default value
+- [fix] rax runApp template export `getSearchParams` method
+
 ## 1.1.0
 
 - [chore] use `enableRouter` replace `buildConfig.router`